Rename `renderpass.rs` to `render_pass.rs`
